api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Ben Jarvis <benjarvis18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
langchain = "^0.2.6"
langchain-openai = "^0.1.14"
azure-cognitiveservices-speech = "^1.38.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
